Impexp_new in Main Mergen.

+++ src/main/java/com/dhbw/thesim/core/entity/Dinosaur.java
@@ -7,6 +7,8 @@ import com.dhbw.thesim.core.util.Vector2D;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Circle;
 
+import java.util.HashMap;
+
 /**
  * Represents a dinosaur object in our simulation.
  * Takes care to update each State and switch between them.
@@ -28,8 +30,8 @@ public class Dinosaur extends SimulationObject {
     private double hydrationFull;
     private double nutrition;
     private double hydration;
-    private int strength;
-    private int speed;
+    private double strength;
+    private double speed;
     private double reproductionRate;
     private double weight;
     private double length;
@@ -67,7 +69,7 @@ public class Dinosaur extends SimulationObject {
      * Constructor for a dinosaur object
      */
     public Dinosaur(String name, String imgName, double nutrition, double hydration,
-                    int strength, int speed, double reproductionRate, double weight, double length, double height,
+                    double strength, double speed, double reproductionRate, double weight, double length, double height,
                     boolean canSwim, boolean canClimb, char diet, double viewRange,
                     double interactionRange, char gender) {
         super(name, interactionRange, "/dinosaur/" + imgName);
@@ -176,11 +178,11 @@ public class Dinosaur extends SimulationObject {
         return hydration;
     }
 
-    public int getStrength() {
+    public double getStrength() {
         return strength;
     }
 
-    public int getSpeed() {
+    public double getSpeed() {
         return speed;
     }
